+++ gazebo/sensors/ForceTorqueSensor_TEST.cc
@@ -19,12 +19,13 @@
 #include "test/ServerFixture.hh"
 #include "test/integration/helper_physics_generator.hh"
 
-
 using namespace gazebo;
+
 class ForceTorqueSensor_TEST : public ServerFixture,
                                public testing::WithParamInterface<const char*>
 {
     public: void ForceTorqueTest(const std::string &_physicsEngine);
+    public: void ForceTorqueTestERB(const std::string &_physicsEngine);
 };
 
 static std::string forceTorqueSensorString =
@@ -39,13 +40,57 @@ static std::string forceTorqueSensorString =
 /// \brief Test Creation of a ForceTorque sensor
 void ForceTorqueSensor_TEST::ForceTorqueTest(const std::string &_physicsEngine)
 {
+  sdf::ElementPtr sdf(new sdf::Element);
+  sdf::initFile("sensor.sdf", sdf);
+  sdf::readString(forceTorqueSensorString, sdf);
+
   Load("worlds/pioneer2dx.world", true, _physicsEngine);
 
   sensors::SensorManager *mgr = sensors::SensorManager::Instance();
 
+  physics::WorldPtr world = physics::get_world("default");
+  physics::ModelPtr model = world->GetModel("pioneer2dx");
+  physics::JointPtr joint = model->GetJoint("left_wheel_hinge");
+
+  // Create the force torque sensor
+  std::string sensorName = mgr->CreateSensor(sdf, "default",
+      "pioneer2dx::left_wheel_hinge", joint->GetId());
+
+  // Make sure the returned sensor name is correct
+  EXPECT_EQ(sensorName,
+      std::string("default::pioneer2dx::left_wheel_hinge::force_torque"));
+
+  // Update the sensor manager so that it can process new sensors.
+  mgr->Update();
+
+  // Get a pointer to the force torque sensor
+  sensors::ForceTorqueSensorPtr sensor =
+    boost::dynamic_pointer_cast<sensors::ForceTorqueSensor>(
+        mgr->GetSensor(sensorName));
+
+  // Make sure the above dynamic cast worked.
+  EXPECT_TRUE(sensor != NULL);
+
+  EXPECT_EQ(sensor->GetTorque(), math::Vector3(0, 0, 0));
+  EXPECT_EQ(sensor->GetForce(), math::Vector3(0, 0, 0));
+
+  EXPECT_TRUE(sensor->IsActive());
+}
+
+/////////////////////////////////////////////////
+/// \brief Test Creation of a ForceTorque sensor, using an ERB file.
+/// The original ForceTorqueTest segfaulted when ERB parsing was introduced.
+/// This test makes sure we don't regress.
+void ForceTorqueSensor_TEST::ForceTorqueTestERB(
+    const std::string &_physicsEngine)
+{
   sdf::ElementPtr sdf(new sdf::Element);
   sdf::initFile("sensor.sdf", sdf);
   sdf::readString(forceTorqueSensorString, sdf);
+
+  Load("worlds/pioneer2dx_erb_test.world", true, _physicsEngine);
+
+  sensors::SensorManager *mgr = sensors::SensorManager::Instance();
 
   physics::WorldPtr world = physics::get_world("default");
   physics::ModelPtr model = world->GetModel("pioneer2dx");
@@ -82,6 +127,13 @@ TEST_P(ForceTorqueSensor_TEST, ForceTorqueTest)
   ForceTorqueTest(GetParam());
 }
 
+/////////////////////////////////////////////////
+TEST_P(ForceTorqueSensor_TEST, ForceTorqueTestERB)
+{
+  ForceTorqueTestERB(GetParam());
+}
+
+/////////////////////////////////////////////////
 INSTANTIATE_TEST_CASE_P(PhysicsEngines,
                         ForceTorqueSensor_TEST,
                         PHYSICS_ENGINE_VALUES);